+++ lib/ReactViews/Custom/Chart/ChartExpandAndDownloadButtons.jsx
@@ -176,6 +176,9 @@ function expandItem(sourceItems, sourceIndex, terria) {
     });
   });
   workbench.add(sourceItem);
+  try {
+    terria.addModel(sourceItem);
+  } catch {}
   runInAction(() =>
     raiseErrorOnRejectedPromise(terria, sourceItem.loadChartItems())
   );