+++ lib/ReactViews/Custom/ChartCustomComponent.ts
@@ -10,8 +10,9 @@ import Chart from "./Chart/FeatureInfoPanelChart";
 import CustomComponent, { ProcessNodeContext } from "./CustomComponent";
 import Model, { BaseModel } from "../../Models/Model";
 import CatalogMemberTraits from "../../Traits/CatalogMemberTraits";
-import CsvCatalogItemTraits from "../../Traits/CsvCatalogItemTraits";
 import hasTraits from "../../Models/hasTraits";
+import SplitItemReference from "../../Models/SplitItemReference";
+import createGuid from "terriajs-cesium/Source/Core/createGuid";
 import DiscretelyTimeVaryingTraits from "../../Traits/DiscretelyTimeVaryingTraits";
 
 export interface ChartCustomComponentAttributes {
@@ -175,6 +176,21 @@ export default abstract class ChartCustomComponent<
     sourceReference: BaseModel | undefined
   ): CatalogItemType;
 
+  /**
+   * For some catalog types, for the chart item to be shareable, it needs to be
+   * constructed as a reference to the original item. This method can be
+   * overriden to make a shareable chart. See SOSChartCustomComponent for an
+   * implementation.
+   *
+   * This method is used only for constructing a chart item to show
+   * in the chart panel, not for the feature info panel chart item.
+   */
+  protected constructShareableCatalogItem?: (
+    id: string | undefined,
+    context: ProcessNodeContext,
+    sourceReference: BaseModel | undefined
+  ) => Promise<CatalogItemType | undefined> = undefined;
+
   private processChart(
     context: ProcessNodeContext,
     node: DomElement,
@@ -190,8 +206,9 @@ export default abstract class ChartCustomComponent<
     const chartDisclaimer = (context.catalogItem as any).chartDisclaimer;
 
     const attrs = this.parseNodeAttrs(node.attribs);
-    const csvString: any =
-      typeof children[0] == "string" ? children[0] : undefined;
+    const child = children[0];
+    const body: string | undefined =
+      typeof child === "string" ? child : undefined;
     const chartElements = [];
     if (!attrs.hideButtons) {
       // Build expand/download buttons
@@ -202,30 +219,32 @@ export default abstract class ChartCustomComponent<
             context.feature.id,
             source
           ].join(":");
-          const item = this.constructCatalogItem(id, context, undefined);
 
-          runInAction(() => {
-            this.setTraitsFromAttrs(item, attrs, i);
+          const itemOrPromise = this.constructShareableCatalogItem
+            ? this.constructShareableCatalogItem(id, context, undefined)
+            : this.constructCatalogItem(id, context, undefined);
 
-            if (
-              csvString &&
-              hasTraits(item, CsvCatalogItemTraits, "csvString")
-            ) {
-              item.setTrait(CommonStrata.user, "csvString", csvString);
+          return Promise.resolve(itemOrPromise).then(item => {
+            if (item) {
+              this.setTraitsFromAttrs(item, attrs, i);
+              body && this.setTraitsFromBody?.(item, body);
+              if (
+                hasTraits(
+                  item,
+                  DiscretelyTimeVaryingTraits,
+                  "chartDisclaimer"
+                ) &&
+                chartDisclaimer !== undefined
+              ) {
+                item.setTrait(
+                  CommonStrata.definition,
+                  "chartDisclaimer",
+                  chartDisclaimer
+                );
+              }
             }
-            if (
-              hasTraits(item, DiscretelyTimeVaryingTraits, "chartDisclaimer") &&
-              chartDisclaimer !== undefined
-            ) {
-              item.setTrait(
-                CommonStrata.definition,
-                "chartDisclaimer",
-                chartDisclaimer
-              );
-            }
+            return item;
           });
-
-          return item;
         }
       );
 
@@ -247,13 +266,7 @@ export default abstract class ChartCustomComponent<
     const chartItem = this.constructCatalogItem(undefined, context, undefined);
     runInAction(() => {
       this.setTraitsFromAttrs(chartItem, attrs, 0);
-
-      if (
-        csvString &&
-        hasTraits(chartItem, CsvCatalogItemTraits, "csvString")
-      ) {
-        chartItem.setTrait(CommonStrata.user, "csvString", csvString);
-      }
+      body && this.setTraitsFromBody?.(chartItem, body);
 
       if (
         hasTraits(chartItem, DiscretelyTimeVaryingTraits, "chartDisclaimer") &&
@@ -302,6 +315,15 @@ export default abstract class ChartCustomComponent<
     attrs: ChartCustomComponentAttributes,
     sourceIndex: number
   ): void;
+
+  /**
+   * Populate  traits in the supplied catalog item with the values from the body of the component.
+   * Assume it will be run in an action.
+   * @param item
+   * @param attrs
+   * @param sourceIndex
+   */
+  protected setTraitsFromBody?: (item: CatalogItemType, body: string) => void;
 
   /**
    * Is this node the first column of a two-column table where the second
@@ -433,6 +455,22 @@ export default abstract class ChartCustomComponent<
       previewXLabel: nodeAttrs["preview-x-label"],
       yColumns
     };
+  }
+
+  /**
+   * A helper method to create a shareable reference to an item.
+   */
+  async createItemReference(
+    sourceItem: CatalogItemType
+  ): Promise<CatalogItemType | undefined> {
+    const terria = sourceItem.terria;
+    const ref = new SplitItemReference(createGuid(), terria);
+    ref.setTrait(CommonStrata.user, "splitSourceItemId", sourceItem.uniqueId);
+    await ref.loadReference();
+    if (ref.target) {
+      terria.addModel(ref);
+      return ref.target as CatalogItemType;
+    }
   }
 }